chore(about): change Constants to Helper
