+++ src/features/about/constants.ts
@@ -1,5 +1,3 @@
-import { IntroInfoSettings } from "./types";
-
 export const TITLE = 'About us';
 
 const PROFILES_LIST = [
@@ -50,18 +48,3 @@ export const PROFILES_REPEAT = [
   ...PROFILES_LIST,
   ...BLANK_LIST,
 ];
-
-export const INTRO_SETTINGS: IntroInfoSettings = {
-  titleOpacity: {
-    startValue: 0,
-    endValue: 1,
-    startScroll: 0,
-    endScroll: 1,
-  },
-  titleLetterSpacing: {
-    startValue: 100,
-    endValue: 0,
-    startScroll: 0,
-    endScroll: 1,
-  },
-};
